+++ src/main/java/ph/devcon/rapidpass/controllers/ExceptionTranslator.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import ph.devcon.rapidpass.services.RegistryService;
 
+import java.io.IOException;
 import java.util.Map;
 
 import static java.util.stream.Collectors.toMap;
@@ -74,5 +75,4 @@ public class ExceptionTranslator {
         log.warn("Server error! ", ex);
         return ImmutableMap.of("message", "Something went wrong! Please contact application owners.");
     }
-
 }